remove conf-libuv from depopts
some ci-scripts try to install depopts, too.  And they fail at this
stage, because libuv1-dev is not available. Therefore libuv is now
always compiled from source.  This has the further advantage that the
included version was properly tested with the current version. And if
the latest version is installed, it's probably newer than the one
shipped by distributions like debian :D